requirements.txt
PyJWT
boto3
python-json-logger
lambdarest
bcrypt
httpx
pyopenssl